+++ merobox/commands/bootstrap/run/executor.py
@@ -2104,6 +2104,25 @@ class WorkflowExecutor:
             "auth_mode": self.auth_mode,
         }
 
+        if step_type in (
+            "account_create",
+            "account_pair",
+            "account_revoke",
+            "account_show",
+        ):
+            from merobox.commands.bootstrap.steps.account import (
+                AccountCreateStep,
+                AccountPairStep,
+                AccountRevokeStep,
+                AccountShowStep,
+            )
+
+            return {
+                "account_create": AccountCreateStep,
+                "account_pair": AccountPairStep,
+                "account_revoke": AccountRevokeStep,
+                "account_show": AccountShowStep,
+            }[step_type](step_config, **common_kwargs)
         if step_type == "stop_node":
             from merobox.commands.bootstrap.steps.stop_node import StopNodeStep